refactor: improve poetry packaging
