+++ .github/workflows/build.yml
@@ -70,7 +70,7 @@ jobs:
             - run: source $VENV
 
             - name: Build Backend
-              run: poetry run pyinstaller --noconfirm ETVR.spec TrackingBackend/main.py
+              run: poetry run pyinstaller --noconfirm ETVR.spec eyetrackvr_backend/main.py
 
             - name: Rename Linux Binary
               if: ${{ matrix.os == 'ubuntu-latest' }}
