+++ .github/workflows/lint.yml
@@ -49,4 +49,4 @@ jobs:
               run: poetry install
             - name: Lint with Mypy
               # Wish we could warn instead of erroring
-              run: poetry run mypy --ignore-missing-imports --check-untyped-defs --show-error-context --implicit-optional --disable-error-code union-attr TrackingBackend/
+              run: poetry run mypy --ignore-missing-imports --check-untyped-defs --show-error-context --implicit-optional --disable-error-code union-attr eyetrackvr_backend/
