--- TrackingBackend/app/__init__.py
@@ -1,4 +1,0 @@
-from .tracker import Tracker
-from .types import EyeData, LogLevel, TrackerPosition
-from .logger import get_logger, setup_logger
-from .config import EyeTrackConfig, CameraConfig, OSCConfig
