+++ eyetrackvr_backend/assets/__init__.py
@@ -1,0 +1,6 @@
+import os.path
+
+from .images import IMAGES_DIR
+from .models import MODELS_DIR
+
+ASSETS_DIR = os.path.dirname(__file__)
