+++ eyetrackvr_backend/assets/images/__init__.py
@@ -1,0 +1,3 @@
+import os.path
+
+IMAGES_DIR = os.path.dirname(__file__)
