+++ eyetrackvr_backend/assets/models/__init__.py
@@ -1,0 +1,3 @@
+import os.path
+
+MODELS_DIR = os.path.dirname(__file__)
