+++ pyproject.toml
@@ -1,12 +1,18 @@
 [tool]
 [tool.poetry]
-name = "EyeTrackVR"
+name = "eyetrackvr_backend"
 version = "1.6.1"
 description = "Opensource, affordable VR eye tracker for VRChat"
 authors = ["ShyAssassin <49711232+ShyAssassin@users.noreply.github.com>", "RedHawk989 <48768484+RedHawk989@users.noreply.github.com>"]
 license = "MIT"
 package-mode = false
 repository = "https://github.com/RedHawk989/EyeTrackVR"
+packages = [
+  { include = "eyetrackvr_backend" }
+]
+
+[tool.poetry.scripts]
+eyetrackvr-backend = 'eyetrackvr_backend.__main__:main'
 
 [tool.poetry.dependencies]
 python = "~3.11.0"
@@ -40,7 +46,7 @@ target-version = ["py310", "py311"]
 
 [tool.ruff]
 select = ["E", "F", "W", "Q"]
-src = ["TrackingBackend", "test"]
+src = ["eyetrackvr-backend", "test"]
 respect-gitignore = true
 target-version = "py311"
 output-format = "grouped"
